requirements.txt
python-dotenv==1.0.0
aiogram==3.2.0
pika-stubs==0.1.3
httpx==0.25.2
aio_pika==9.3.1